Separar las funciones utiles del programa / Ponerlas en un archivo
