+++ .gitignore
@@ -1,2 +1,6 @@
 __pycache__
 *.pyc
+.venv
+.env
+*.sql
+.nicegui
