+++ README.md
@@ -1,1 +1,9 @@
-# Eye_System-Definitivo
+# Eye_System-Core
+
+## Trabajo en proceso...
+### Para hacer.
+- Interfaz Web
+- Soporte SQLite
+- Dockerfile
+- Cronjob de entrenamiento
+- Ampliar datasets de entrenamiento en base a el reconocimiento actual
