--- formularios/__init__.py
@@ -1,2 +1,0 @@
-# __init__.py
-__all__ = []
