--- formularios/config.py
@@ -1,8 +1,0 @@
-COLOR_BARRA_SUPERIOR = "#1f2329"
-COLOR_MENU_LATERAL = "#2a3138"
-COLOR_CUERPO_PRINCIPAL = "#f1faff"
-COLOR_MENU_CURSOR_ENCIMA = "#2f88c5"
-COLOR_PRINCIPAL = "#1a1b26"     # Azul oscuro
-COLOR_SECUNDARIO = "#f0f5ff"    # Azul claro
-COLOR_ACENTO = "#2d4f7c"        # Azul medio
-COLOR_HOVER = "#1a365d"         # Azul oscuro para hover
