+++ src/__init__.py
@@ -1,0 +1,2 @@
+# __init__.py
+__all__ = []
